+++ hpOneView/resources/networking/connection_templates.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 ###
-# (C) Copyright (2012-2017) Hewlett Packard Enterprise Development LP
+# (C) Copyright (2012-2019) Hewlett Packard Enterprise Development LP
 #
 # Permission is hereby granted, free of charge, to any person obtaining a copy
 # of this software and associated documentation files (the "Software"), to deal
@@ -31,10 +31,10 @@ from future import standard_library
 standard_library.install_aliases()
 
 
-from hpOneView.resources.resource import ResourceClient
+from hpOneView.resources.resource import Resource
 
 
-class ConnectionTemplates(object):
+class ConnectionTemplates(Resource):
     """
     Connection Templates API client.
 
@@ -44,67 +44,13 @@ class ConnectionTemplates(object):
     DEFAULT_VALUES = {
         '200': {"type": "connection-template"},
         '300': {"type": "connection-template"},
-        '500': {"type": "connection-template"}
+        '500': {"type": "connection-template"},
+        '600': {"type": "connection-template"},
+        '800': {"type": "connection-template"}
     }
 
-    def __init__(self, con):
-        self._connection = con
-        self._client = ResourceClient(con, self.URI)
-
-    def get_all(self, start=0, count=-1, filter='', sort=''):
-        """
-        Gets a paginated collection of all connection templates based on the specified
-        parameters. Filters can be used in the URL to control the number of connection
-        templates that are returned.
-
-        With no filters specified, the API returns all connection templates.
-
-        Args:
-            start:
-                The first item to return, using 0-based indexing.
-                If not specified, the default is 0 - start with the first available item.
-            count:
-                The number of resources to return. A count of -1 requests all items.
-                The actual number of items in the response might differ from the requested
-                count if the sum of start and count exceeds the total number of items.
-            filter (list or str):
-                A general filter/query string to narrow the list of items returned. The
-                default is no filter; all resources are returned.
-            sort:
-                The sort order of the returned data set. By default, the sort order is based
-                on create time with the oldest entry first.
-
-        Returns:
-            list: A list of connection templates.
-        """
-        return self._client.get_all(start, count, filter=filter, sort=sort)
-
-    def get(self, id_or_uri):
-        """
-        Gets the connection template with the specified ID or URI.
-
-        Args:
-            id_or_uri: ID or URI of connection template.
-
-        Returns:
-            dict: the connection template
-        """
-        return self._client.get(id_or_uri)
-
-    def get_by(self, field, value):
-        """
-        Gets all connection templates that match the filter.
-
-        The search is case-insensitive.
-
-        Args:
-            field: Field name to filter.
-            value: Value to filter.
-
-        Returns:
-            list: A list of connection templates.
-        """
-        return self._client.get_by(field, value)
+    def __init__(self, connection, data=None):
+        super(ConnectionTemplates, self).__init__(connection, data)
 
     def get_default(self):
         """
@@ -115,19 +61,4 @@ class ConnectionTemplates(object):
             dict:
         """
         uri = self.URI + "/defaultConnectionTemplate"
-        return self._client.get(uri)
-
-    def update(self, resource, timeout=-1):
-        """
-        Updates connection template.
-
-        Args:
-            resource (dict): Object to update.
-            timeout:
-                Timeout in seconds. Wait for task completion by default. The timeout does not abort the operation in
-                OneView, just stops waiting for its completion.
-
-        Returns:
-            dict: Updated resource.
-        """
-        return self._client.update(resource, timeout=timeout, default_values=self.DEFAULT_VALUES)
+        return self._helper.do_get(uri)